Suppressed some warnings

+++ idk/projects/vkn/VulkanMeshFactory.cpp
@@ -208,9 +208,9 @@ namespace idk::vkn
 					// Subdivide every face once
 					for (unsigned j = 0; j < tmpIndices.size();)
 					{
-						const unsigned v1Index = tmpIndices[j++];
-						const unsigned v2Index = tmpIndices[j++];
-						const unsigned v3Index = tmpIndices[j++];
+						const uint16_t v1Index = tmpIndices[j++];
+						const uint16_t v2Index = tmpIndices[j++];
+						const uint16_t v3Index = tmpIndices[j++];
 
 						const vec3 v1 = vertices[v1Index].pos;
 						const vec3 v2 = vertices[v2Index].pos;
@@ -221,15 +221,15 @@ namespace idk::vkn
 						// addSubVert will check if the vertex is unique or not using the sharedIndices set.
 						// v1 -> v2
 						const vec3 hv1 = (v1 + v2).normalize();
-						const unsigned hv1Index = addSubVert(vertices, sharedIndices, hv1);
+						const uint16_t hv1Index = addSubVert(vertices, sharedIndices, hv1);
 
 						// v2 -> v3
 						const vec3 hv2 = (v2 + v3).normalize();
-						const unsigned hv2Index = addSubVert(vertices, sharedIndices, hv2);
+						const uint16_t hv2Index = addSubVert(vertices, sharedIndices, hv2);
 
 						// v3 -> v1
 						const vec3 hv3 = (v3 + v1).normalize();
-						const unsigned hv3Index = addSubVert(vertices, sharedIndices, hv3);
+						const uint16_t hv3Index = addSubVert(vertices, sharedIndices, hv3);
 
 						// New Indices. Every triangle will subdivide into 4 new triangles.
 						indices.push_back(v1Index);
